pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "AndreyFinyak",email = "afinyak@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic (>=1.17.1,<2.0.0)",
    "fastapi[standard] (>=0.121.0,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "dependency-injector (>=4.48.2,<5.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pre-commit (>=4.4.0,<5.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "types-aiofiles (>=25.1.0.20251011,<26.0.0.0)"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
